fix: ssr errors for showcase; broken links
