+++ docs/fundamentals/README.md
@@ -35,4 +35,4 @@ import DocCardList from '@theme/DocCardList';
 
 <DocCardList />
 
-Be sure to also review the [Tableland SQL guides](/sql), which describe how to use SQL and details some web3 constraints that are important to understand when writing queries. If you're new to Tableland, there are a lot terms used throughout the protocol and documentation. The [glossary](/fundamentals/about/glossary) provides a description of various terms, and the [general FAQs](/fundamentals/about/general-faqs) are high-level common questions & answers. Note that more specific FAQs are scattered about the docs that are located within a specific topic's documentation.
+Be sure to also review the [Tableland SQL guides](/sql), which describe how to use SQL and details some web3 constraints that are important to understand when writing queries. If you're new to Tableland, there are a lot terms used throughout the protocol and documentation. The [glossary](/fundamentals/about/glossary) provides a description of various terms, and the [general FAQs](/fundamentals/about/faqs) are high-level common questions & answers. Note that more specific FAQs are scattered about the docs that are located within a specific topic's documentation.
