+++ src/components/projects.ts
@@ -1,40 +1,8 @@
-export interface Project {
-  name: string; // Name of the project
-  description: string; // Short description
-  logo: string; // Logo filename in `src/static/img/showcase`
-  protocol: Protocol; // Either Tableland or Textile
-  tags?: Tags[]; // Tags for the project, for filtering
-  chains?: Chains[]; // Chains the project is on, for filtering
-  website: string; // Website link
-  github?: string; // Github link
-  twitter: string; // Twitter link
-  details: string; // Long form details about the project
-  youtubeId?: string; // Youtube demo ID, e.g., `-MUq--Nrd0c` for `https://www.youtube.com/watch?v=-MUq--Nrd0c`
-}
+// Note: Docusaurus expects all files in `/pages` to export a React component,
+// so this file must be moved elsewhere to avoid compilation errors.
+import { type Project } from "../pages/showcase";
 
-export type Protocol = "Tableland" | "Textile";
-
-export type Tags =
-  | "NFT"
-  | "Gaming"
-  | "DAO"
-  | "DeFi"
-  | "Infra"
-  | "Social"
-  | "Dataset"
-  | "DePIN"
-  | "dApp"
-  | "AI"
-  | "ML";
-
-export type Chains =
-  | "Ethereum"
-  | "Filecoin"
-  | "Arbitrum"
-  | "Optimism"
-  | "Polygon";
-
-export const projects = [
+export default [
   {
     name: "DIMO",
     description:
